requirements.txt
streamlit
openpyxl
pandas
matplotlib
python-docx
XlsxWriter